+++ web/ASC.Web.Client/public/index.html
@@ -43,174 +43,12 @@
         font-size: 13px;
         -webkit-font-smoothing: antialiased;
       }
-
-      .temp-header-container {
-        align-items: center;
-        background-color: rgb(15, 64, 113);
-        display: grid;
-        grid-template-columns: 24px 168px 1fr 36px;
-        grid-template-rows: 1fr;
-        grid-column-gap: 16px;
-        width: calc(100vw - 16px);
-        height: 56px;
-        padding-right: 16px;
-      }
-      .temp-content-loader {
-        padding: 16px;
-        height: calc(100vh - 91px);
-        display: grid;
-        grid-template-columns: 1fr;
-        grid-template-rows: calc(100vh - 91px);
-        grid-column-gap: 8px;
-      }
-      @media (max-width: 1024px) {
-        .temp-content-loader {
-          grid-template-columns: 1fr;
-        }
-        .temp-content__article {
-          display: none;
-        }
-      }
-      .burger-loader-svg {
-        display: none;
-        width: 24px;
-        height: 24px;
-        padding-left: 16px;
-      }
-      .logo-loader-svg {
-        display: none;
-        width: 168px;
-        height: 24px;
-        position: relative;
-        cursor: pointer;
-        padding-left: 16px;
-      }
-      .avatar-loader-svg {
-        display: none;
-        width: 36px;
-        height: 36px;
-      }
     </style>
     <title><%= htmlWebpackPlugin.options.title %></title>
   </head>
   <body>
     <noscript> You need to enable JavaScript to run this app. </noscript>
     <div class="ipl-progress-indicator" id="ipl-progress-indicator"></div>
-    <div id="temp-content" style="display: none;">
-      <header class="temp-header-container">
-        <div id="burger-loader-svg" class="burger-loader-svg">
-          <svg
-            role="img"
-            width="100%"
-            height="100%"
-            aria-labelledby="loading-aria0"
-            preserveAspectRatio="none"
-          >
-            <rect
-              rx="3"
-              ry="3"
-              width="24"
-              height="24"
-              style="fill: url('#fill0');"
-            ></rect>
-            <defs>
-              <linearGradient id="fill0">
-                <stop
-                  offset="0.599964"
-                  stop-color="#fff"
-                  stop-opacity="0.25"
-                ></stop>
-              </linearGradient>
-            </defs>
-          </svg>
-        </div>
-        <div id="logo-loader-svg" class="logo-loader-svg">
-          <svg
-            role="img"
-            width="100%"
-            height="100%"
-            aria-labelledby="loading-aria01"
-            preserveAspectRatio="none"
-          >
-            <rect
-              rx="3"
-              ry="3"
-              width="168"
-              height="24"
-              style="fill: url('#fill01');"
-            ></rect>
-            <defs>
-              <linearGradient id="fill01">
-                <stop
-                  offset="0.599964"
-                  stop-color="#fff"
-                  stop-opacity="0.25"
-                ></stop>
-              </linearGradient>
-            </defs>
-          </svg>
-        </div>
-        <div></div>
-        <div id="avatar-loader-svg" class="avatar-loader-svg">
-          <svg
-            role="img"
-            width="100%"
-            height="100%"
-            aria-labelledby="loading-aria01"
-            preserveAspectRatio="none"
-          >
-            <circle
-              cx="18"
-              cy="18"
-              r="18"
-              width="36"
-              height="36"
-              style="fill: url('#fill02');"
-            ></circle>
-            <defs>
-              <linearGradient id="fill02">
-                <stop
-                  offset="0.599964"
-                  stop-color="#fff"
-                  stop-opacity="0.25"
-                ></stop>
-              </linearGradient>
-            </defs>
-          </svg>
-        </div>
-      </header>
-
-      <div class="temp-content-loader">
-        <svg
-          role="img"
-          width="100%"
-          height="100%"
-          aria-labelledby="loading-aria1"
-          preserveAspectRatio="none"
-        >
-          <rect
-            x="0"
-            y="0"
-            width="100%"
-            height="100%"
-            clip-path="url(#clip-path1)"
-            style="fill: url('#fill1');"
-          ></rect>
-          <defs>
-            <clipPath id="clip-path1">
-              <rect x="3" y="3" rx="5" ry="5" width="100%" />
-            </clipPath>
-            <linearGradient id="fill1">
-              <stop
-                offset="0.599964"
-                stop-color="#000000"
-                stop-opacity="0.1"
-              ></stop>
-            </linearGradient>
-          </defs>
-        </svg>
-      </div>
-    </div>
 
     <div id="root"></div>
     <!--
@@ -225,16 +63,6 @@
     -->
     <script>
       console.log("It's WEB CLIENT INIT");
-      console.log("AscDesktopEditor", window["AscDesktopEditor"]);
-
-      var pathname = window.location.pathname.toLowerCase();
-      if (
-        pathname.indexOf("thirdparty") === -1 &&
-        window["AscDesktopEditor"] === undefined
-      ) {
-        let el = document.getElementById("temp-content");
-        el.style.display = "block";
-      }
     </script>
   </body>
 </html>
